refactor(frontend): Phase 4 medium-priority component extraction
Extract components from 4 medium-priority files to improve maintainability:

CustomFormatMapper.tsx (839→468 lines, 44% reduction)
- types/customFormat.ts: shared types and constants
- components/format-mapper/ConfidenceDisplay.tsx: confidence UI
- components/format-mapper/PreviewTable.tsx: transaction preview
- hooks/useFormatDetection.ts: auto-detection logic

import/page.tsx (793→353 lines, 55% reduction)
- types/import.ts: import types and FORMAT_NAMES
- components/import/BatchImport.tsx: batch import UI
- components/import/SingleFileImport.tsx: single file import UI
- components/import/ImportResult.tsx: result display

admin/page.tsx (762→358 lines, 53% reduction)
- types/admin.ts: admin types and TAG_TABS
- components/admin/OverviewTab.tsx: stats overview
- components/admin/ImportsTab.tsx: import sessions table
- components/admin/TagsTab.tsx: tag management table
- components/admin/TagModal.tsx: create/edit tag modals

dashboard/configure/page.tsx (709→316 lines, 55% reduction)
- types/dashboard.ts: widget types and WIDGET_INFO
- components/dashboard/WidgetRow.tsx: widget row component
- components/dashboard/WidgetEditModal.tsx: edit modal
- components/dashboard/WidgetReference.tsx: widget reference grid

Total: 3,103 lines reduced to 1,495 lines (52% reduction)
