+++ frontend/src/app/import/page.tsx
@@ -1,57 +1,23 @@
 'use client'
 
 import { useState, useEffect } from 'react'
-import { format } from 'date-fns'
-import { formatCurrency } from '@/lib/format'
 import { PageHelp } from '@/components/PageHelp'
-
-interface AccountTag {
-  id: number
-  namespace: string
-  value: string
-}
-
-interface SavedCustomFormat {
-  id: number
-  name: string
-  description: string | null
-  config_json: string
-  use_count: number
-}
-
-const FORMAT_NAMES: Record<string, string> = {
-  'bofa_bank': 'Bank of America (Checking/Savings)',
-  'bofa_cc': 'Bank of America (Credit Card)',
-  'amex_cc': 'American Express',
-  'inspira_hsa': 'Inspira HSA',
-  'venmo': 'Venmo',
-  'qif': 'Quicken (QIF)',
-  'qfx': 'Quicken/OFX (QFX)',
-  'custom': 'Custom Format',
-  'unknown': 'Unknown'
-}
-
-interface FilePreview {
-  filename: string
-  account_source: string | null
-  detected_format: string
-  transaction_count: number
-  duplicate_count: number
-  cross_file_duplicate_count: number
-  total_amount: number
-  date_range_start: string | null
-  date_range_end: string | null
-  transactions: any[]
-  selected: boolean
-  accountSourceOverride?: string
-}
+import { BatchImport } from '@/components/import/BatchImport'
+import { SingleFileImport } from '@/components/import/SingleFileImport'
+import { ImportResult } from '@/components/import/ImportResult'
+import {
+  AccountTag,
+  SavedCustomFormat,
+  FilePreview,
+  ImportResult as ImportResultType
+} from '@/types/import'
 
 export default function ImportPage() {
   const [files, setFiles] = useState<File[]>([])
   const [batchMode, setBatchMode] = useState(false)
   const [batchPreviews, setBatchPreviews] = useState<FilePreview[]>([])
   const [importing, setImporting] = useState(false)
-  const [result, setResult] = useState<any>(null)
+  const [result, setResult] = useState<ImportResultType | null>(null)
 
   // Account selection state
   const [accounts, setAccounts] = useState<AccountTag[]>([])
@@ -156,7 +122,7 @@ export default function ImportPage() {
           ...config,
           account_source: accountSource || config.account_source
         }))
-        formData.append('save_config', 'false') // Already saved
+        formData.append('save_config', 'false')
 
         const res = await fetch('/api/v1/import/custom/confirm', {
           method: 'POST',
@@ -344,450 +310,44 @@ export default function ImportPage() {
 
       {/* Batch Import UI */}
       {batchMode && (
-        <>
-          <div className="bg-white rounded-lg shadow p-6 space-y-4">
-            <div>
-              <label className="block text-sm font-medium text-gray-700 mb-2">
-                Import Files (Multiple)
-              </label>
-              <input
-                type="file"
-                accept=".csv,.qif,.qfx,.ofx"
-                multiple
-                onChange={(e) => setFiles(Array.from(e.target.files || []))}
-                className="block w-full text-sm text-gray-500 file:mr-4 file:py-2 file:px-4 file:rounded file:border-0 file:text-sm file:font-semibold file:bg-blue-50 file:text-blue-700 hover:file:bg-blue-100"
-              />
-              {files.length > 0 && (
-                <p className="mt-2 text-sm text-gray-600">
-                  {files.length} file{files.length !== 1 ? 's' : ''} selected
-                </p>
-              )}
-            </div>
-
-            <button
-              onClick={handleBatchPreview}
-              disabled={files.length === 0}
-              className="w-full px-4 py-2 bg-blue-600 text-white rounded-md hover:bg-blue-700 disabled:bg-gray-300 disabled:cursor-not-allowed"
-            >
-              Preview Batch Import
-            </button>
-          </div>
-
-          {/* Batch Previews */}
-          {batchPreviews.length > 0 && (
-            <div className="bg-white rounded-lg shadow p-6 space-y-6">
-              <div className="flex justify-between items-center">
-                <h2 className="text-xl font-semibold">Batch Import Preview</h2>
-                <div className="text-sm text-gray-600">
-                  {batchPreviews.filter(p => p.selected).length} of {batchPreviews.length} files selected
-                </div>
-              </div>
-
-              {/* Summary Stats */}
-              <div className="grid grid-cols-3 gap-4 p-4 bg-gray-50 rounded">
-                <div>
-                  <p className="text-sm text-gray-600">Total Transactions</p>
-                  <p className="text-2xl font-bold">
-                    {batchPreviews.filter(p => p.selected).reduce((sum, p) => sum + p.transaction_count, 0)}
-                  </p>
-                </div>
-                <div>
-                  <p className="text-sm text-gray-600">Total Duplicates</p>
-                  <p className="text-2xl font-bold text-yellow-600">
-                    {batchPreviews.filter(p => p.selected).reduce((sum, p) => sum + p.duplicate_count + p.cross_file_duplicate_count, 0)}
-                  </p>
-                </div>
-                <div>
-                  <p className="text-sm text-gray-600">Net Amount</p>
-                  <p className={`text-2xl font-bold ${batchPreviews.filter(p => p.selected).reduce((sum, p) => sum + p.total_amount, 0) >= 0 ? 'text-green-600' : 'text-red-600'}`}>
-                    {formatCurrency(batchPreviews.filter(p => p.selected).reduce((sum, p) => sum + p.total_amount, 0))}
-                  </p>
-                </div>
-              </div>
-
-              {/* File List */}
-              <div className="space-y-4">
-                {batchPreviews.map((filePreview) => (
-                  <div key={filePreview.filename} className={`border rounded-lg p-4 ${filePreview.selected ? 'border-blue-500 bg-blue-50' : 'border-gray-200'}`}>
-                    <div className="flex items-start gap-4">
-                      <input
-                        type="checkbox"
-                        checked={filePreview.selected}
-                        onChange={() => toggleFileSelection(filePreview.filename)}
-                        className="mt-1"
-                      />
-                      <div className="flex-1 space-y-3">
-                        <div className="flex justify-between items-start">
-                          <div>
-                            <h3 className="font-semibold text-lg">{filePreview.filename}</h3>
-                            <span className="px-2 py-1 bg-gray-200 text-gray-700 rounded text-xs">
-                              {FORMAT_NAMES[filePreview.detected_format] || filePreview.detected_format}
-                            </span>
-                          </div>
-                        </div>
-
-                        <div className="grid grid-cols-5 gap-3 text-sm">
-                          <div>
-                            <p className="text-gray-600">Transactions</p>
-                            <p className="font-semibold">{filePreview.transaction_count}</p>
-                          </div>
-                          <div>
-                            <p className="text-gray-600">DB Duplicates</p>
-                            <p className="font-semibold text-yellow-600">{filePreview.duplicate_count}</p>
-                          </div>
-                          <div>
-                            <p className="text-gray-600">Batch Duplicates</p>
-                            <p className="font-semibold text-orange-600">{filePreview.cross_file_duplicate_count}</p>
-                          </div>
-                          <div>
-                            <p className="text-gray-600">Date Range</p>
-                            <p className="font-semibold">
-                              {filePreview.date_range_start && filePreview.date_range_end
-                                ? `${format(new Date(filePreview.date_range_start), 'MM/dd/yy')} - ${format(new Date(filePreview.date_range_end), 'MM/dd/yy')}`
-                                : 'N/A'}
-                            </p>
-                          </div>
-                          <div>
-                            <p className="text-gray-600">Total Amount</p>
-                            <p className={`font-semibold ${filePreview.total_amount >= 0 ? 'text-green-600' : 'text-red-600'}`}>
-                              {formatCurrency(filePreview.total_amount)}
-                            </p>
-                          </div>
-                        </div>
-
-                        <div>
-                          <label className="block text-xs text-gray-600 mb-1">
-                            Account Source <span className="text-red-500">*</span>
-                          </label>
-                          <div className="flex gap-2">
-                            {accounts.length > 0 && (
-                              <select
-                                value={filePreview.accountSourceOverride || filePreview.account_source || ''}
-                                onChange={(e) => updateAccountSource(filePreview.filename, e.target.value)}
-                                className={`flex-1 px-3 py-1 text-sm border rounded-md ${!(filePreview.accountSourceOverride || filePreview.account_source) ? 'border-yellow-400' : ''}`}
-                              >
-                                <option value="">-- Select Account --</option>
-                                {accounts.map((acct) => (
-                                  <option key={acct.id} value={acct.value}>
-                                    {acct.value}
-                                  </option>
-                                ))}
-                              </select>
-                            )}
-                            <input
-                              type="text"
-                              value={filePreview.accountSourceOverride || filePreview.account_source || ''}
-                              onChange={(e) => updateAccountSource(filePreview.filename, e.target.value)}
-                              placeholder={accounts.length > 0 ? "or type new" : "e.g., BOFA-Checking"}
-                              className={`${accounts.length > 0 ? 'w-32' : 'flex-1'} px-3 py-1 text-sm border rounded-md ${!(filePreview.accountSourceOverride || filePreview.account_source) ? 'border-yellow-400' : ''}`}
-                            />
-                          </div>
-                          {!(filePreview.accountSourceOverride || filePreview.account_source) && (
-                            <p className="text-xs text-yellow-600 mt-1">Required for duplicate detection</p>
-                          )}
-                        </div>
-                      </div>
-                    </div>
-                  </div>
-                ))}
-              </div>
-
-              <div className="flex gap-4">
-                {(() => {
-                  const selectedFiles = batchPreviews.filter(p => p.selected)
-                  const filesWithoutAccount = selectedFiles.filter(p => !(p.accountSourceOverride || p.account_source))
-                  const allHaveAccounts = filesWithoutAccount.length === 0
-                  const isDisabled = importing || selectedFiles.length === 0 || !allHaveAccounts
-
-                  return (
-                    <button
-                      onClick={handleBatchConfirm}
-                      disabled={isDisabled}
-                      className="flex-1 px-4 py-2 bg-green-600 text-white rounded-md hover:bg-green-700 disabled:bg-gray-300 disabled:cursor-not-allowed"
-                    >
-                      {importing ? 'Importing...' :
-                       !allHaveAccounts ? `${filesWithoutAccount.length} File${filesWithoutAccount.length !== 1 ? 's' : ''} Missing Account` :
-                       `Import ${selectedFiles.length} Selected File${selectedFiles.length !== 1 ? 's' : ''}`}
-                    </button>
-                  )
-                })()}
-                <button
-                  onClick={() => setBatchPreviews([])}
-                  className="px-4 py-2 bg-gray-200 text-gray-700 rounded-md hover:bg-gray-300"
-                >
-                  Cancel
-                </button>
-              </div>
-            </div>
-          )}
-        </>
+        <BatchImport
+          files={files}
+          setFiles={setFiles}
+          batchPreviews={batchPreviews}
+          accounts={accounts}
+          importing={importing}
+          onPreview={handleBatchPreview}
+          onConfirm={handleBatchConfirm}
+          onCancel={() => setBatchPreviews([])}
+          onToggleSelection={toggleFileSelection}
+          onUpdateAccountSource={updateAccountSource}
+        />
       )}
 
       {/* Single File Import UI */}
       {!batchMode && (
-        <>
-          <div className="bg-white rounded-lg shadow p-6 space-y-4">
-            <div>
-              <label className="block text-sm font-medium text-gray-700 mb-2">
-                Import File
-              </label>
-              <input
-                type="file"
-                accept=".csv,.qif,.qfx,.ofx"
-                onChange={(e) => setFile(e.target.files?.[0] || null)}
-                className="block w-full text-sm text-gray-500 file:mr-4 file:py-2 file:px-4 file:rounded file:border-0 file:text-sm file:font-semibold file:bg-blue-50 file:text-blue-700 hover:file:bg-blue-100"
-              />
-            </div>
-
-            <div className="grid grid-cols-1 md:grid-cols-2 gap-4">
-              <div>
-                <div className="flex items-center justify-between mb-2">
-                  <label className="text-sm font-medium text-gray-700">
-                    Account Source <span className="text-red-500">*</span>
-                  </label>
-                  {accounts.length > 0 && (
-                    <div className="flex gap-1">
-                      <button
-                        type="button"
-                        onClick={() => {
-                          setAccountMode('existing')
-                          setAccountSource('')
-                        }}
-                        className={`px-2 py-0.5 text-xs rounded ${accountMode === 'existing' ? 'bg-blue-600 text-white' : 'bg-gray-200 text-gray-700'}`}
-                      >
-                        Existing
-                      </button>
-                      <button
-                        type="button"
-                        onClick={() => {
-                          setAccountMode('new')
-                          setAccountSource('')
-                        }}
-                        className={`px-2 py-0.5 text-xs rounded ${accountMode === 'new' ? 'bg-blue-600 text-white' : 'bg-gray-200 text-gray-700'}`}
-                      >
-                        New
-                      </button>
-                    </div>
-                  )}
-                </div>
-                {accountMode === 'existing' && accounts.length > 0 ? (
-                  <select
-                    value={accountSource}
-                    onChange={(e) => setAccountSource(e.target.value)}
-                    className={`w-full px-4 py-2 border rounded-md ${!accountSource ? 'border-yellow-400' : ''}`}
-                  >
-                    <option value="">-- Select Account --</option>
-                    {accounts.map((acct) => (
-                      <option key={acct.id} value={acct.value}>
-                        {acct.value}
-                      </option>
-                    ))}
-                  </select>
-                ) : (
-                  <input
-                    type="text"
-                    placeholder="e.g., BOFA-Checking, AMEX-53004"
-                    value={accountSource}
-                    onChange={(e) => setAccountSource(e.target.value)}
-                    className={`w-full px-4 py-2 border rounded-md ${!accountSource ? 'border-yellow-400' : ''}`}
-                  />
-                )}
-                {!accountSource && (
-                  <p className="text-xs text-yellow-600 mt-1">
-                    Required for accurate duplicate detection
-                  </p>
-                )}
-              </div>
-
-              <div>
-                <label className="block text-sm font-medium text-gray-700 mb-2">
-                  Format (Optional)
-                </label>
-                <select
-                  value={selectedCustomFormat ? `custom:${selectedCustomFormat.id}` : formatHint}
-                  onChange={(e) => handleFormatChange(e.target.value)}
-                  className="w-full px-4 py-2 border rounded-md"
-                >
-                  <option value="">Auto-detect</option>
-                  <optgroup label="Standard File Formats">
-                    <option value="qif">Quicken (QIF)</option>
-                    <option value="qfx">Quicken/OFX (QFX)</option>
-                  </optgroup>
-                  {savedFormats.length > 0 && (
-                    <optgroup label="Saved Custom Formats">
-                      {savedFormats.map(fmt => (
-                        <option key={fmt.id} value={`custom:${fmt.id}`}>
-                          {fmt.name}
-                        </option>
-                      ))}
-                    </optgroup>
-                  )}
-                </select>
-                {selectedCustomFormat && (
-                  <p className="mt-1 text-xs text-purple-600">
-                    Using saved format: {selectedCustomFormat.name}
-                  </p>
-                )}
-              </div>
-            </div>
-
-            <button
-              onClick={handlePreview}
-              disabled={!file || !accountSource}
-              className="w-full px-4 py-2 bg-blue-600 text-white rounded-md hover:bg-blue-700 disabled:bg-gray-300 disabled:cursor-not-allowed"
-            >
-              {!accountSource ? 'Select an Account to Preview' : 'Preview Import'}
-            </button>
-          </div>
-
-          {/* Preview */}
-          {preview && (
-            <div className="bg-white rounded-lg shadow p-6 space-y-4">
-              <div className="flex justify-between items-center">
-                <h2 className="text-xl font-semibold">Preview</h2>
-                <span className="px-3 py-1 bg-blue-100 text-blue-800 rounded-full text-sm">
-                  {selectedCustomFormat
-                    ? selectedCustomFormat.name
-                    : FORMAT_NAMES[preview.detected_format] || preview.detected_format}
-                </span>
-              </div>
-
-              {preview.errors && preview.errors.length > 0 && (
-                <div className="p-3 bg-red-50 border border-red-200 rounded">
-                  <p className="text-sm font-medium text-red-800">Parsing Errors:</p>
-                  <ul className="mt-1 text-sm text-red-600 list-disc list-inside">
-                    {preview.errors.slice(0, 3).map((err: string, idx: number) => (
-                      <li key={idx}>{err}</li>
-                    ))}
-                  </ul>
-                </div>
-              )}
-
-              <div className="grid grid-cols-2 gap-4 p-4 bg-gray-50 rounded">
-                <div>
-                  <p className="text-sm text-gray-600">Transactions</p>
-                  <p className="text-2xl font-bold">{preview.transaction_count}</p>
-                </div>
-                <div>
-                  <p className="text-sm text-gray-600">Total Amount</p>
-                  <p className={`text-2xl font-bold ${preview.total_amount >= 0 ? 'text-green-600' : 'text-red-600'}`}>
-                    {formatCurrency(preview.total_amount)}
-                  </p>
-                </div>
-              </div>
-
-              <div className="overflow-x-auto">
-                <table className="min-w-full divide-y divide-gray-200">
-                  <thead className="bg-gray-50">
-                    <tr>
-                      <th className="px-4 py-2 text-left text-xs font-medium text-gray-500 uppercase">Date</th>
-                      <th className="px-4 py-2 text-left text-xs font-medium text-gray-500 uppercase">Merchant</th>
-                      <th className="px-4 py-2 text-left text-xs font-medium text-gray-500 uppercase">Bucket</th>
-                      <th className="px-4 py-2 text-right text-xs font-medium text-gray-500 uppercase">Amount</th>
-                    </tr>
-                  </thead>
-                  <tbody className="bg-white divide-y divide-gray-200">
-                    {preview.transactions.slice(0, 10).map((txn: any, idx: number) => (
-                      <tr key={idx}>
-                        <td className="px-4 py-2 text-sm">{format(new Date(txn.date), 'MM/dd/yyyy')}</td>
-                        <td className="px-4 py-2 text-sm">{txn.merchant}</td>
-                        <td className="px-4 py-2 text-sm">
-                          <span className="px-2 py-1 bg-gray-100 rounded text-xs">{txn.bucket || 'No Bucket'}</span>
-                        </td>
-                        <td className={`px-4 py-2 text-sm text-right ${txn.amount >= 0 ? 'text-green-600' : 'text-red-600'}`}>
-                          {formatCurrency(txn.amount, true)}
-                        </td>
-                      </tr>
-                    ))}
-                  </tbody>
-                </table>
-                {preview.transactions.length > 10 && (
-                  <p className="text-center text-sm text-gray-500 mt-4">
-                    Showing first 10 of {preview.transaction_count} transactions
-                  </p>
-                )}
-              </div>
-
-              <div className="flex gap-4">
-                <button
-                  onClick={handleConfirm}
-                  disabled={importing}
-                  className="flex-1 px-4 py-2 bg-green-600 text-white rounded-md hover:bg-green-700 disabled:bg-gray-300 disabled:cursor-not-allowed"
-                >
-                  {importing ? 'Importing...' : 'Confirm Import'}
-                </button>
-                <button
-                  onClick={() => setPreview(null)}
-                  className="px-4 py-2 bg-gray-200 text-gray-700 rounded-md hover:bg-gray-300"
-                >
-                  Cancel
-                </button>
-              </div>
-            </div>
-          )}
-        </>
+        <SingleFileImport
+          file={file}
+          setFile={setFile}
+          accounts={accounts}
+          accountSource={accountSource}
+          setAccountSource={setAccountSource}
+          accountMode={accountMode}
+          setAccountMode={setAccountMode}
+          formatHint={formatHint}
+          savedFormats={savedFormats}
+          selectedCustomFormat={selectedCustomFormat}
+          preview={preview}
+          importing={importing}
+          onPreview={handlePreview}
+          onConfirm={handleConfirm}
+          onCancelPreview={() => setPreview(null)}
+          onFormatChange={handleFormatChange}
+        />
       )}
 
       {/* Result */}
-      {result && (
-        <div className="bg-green-50 border border-green-200 rounded-lg p-6 space-y-4">
-          <h2 className="text-xl font-semibold text-green-900">Import Complete!</h2>
-
-          <div className="grid grid-cols-3 gap-4">
-            <div>
-              <p className="text-sm text-green-700">Imported</p>
-              <p className="text-2xl font-bold text-green-900">
-                {result.total_imported || result.imported}
-              </p>
-            </div>
-            <div>
-              <p className="text-sm text-green-700">Duplicates Skipped</p>
-              <p className="text-2xl font-bold text-green-900">
-                {result.total_duplicates || result.duplicates}
-              </p>
-            </div>
-            <div>
-              <p className="text-sm text-green-700">
-                {result.files ? 'Files Imported' : 'Format Saved'}
-              </p>
-              <p className="text-2xl font-bold text-green-900">
-                {result.files ? result.files.length : (result.format_saved || result.config_saved ? 'Yes' : 'No')}
-              </p>
-            </div>
-          </div>
-
-          {/* Per-file results for batch imports */}
-          {result.files && result.files.length > 0 && (
-            <div className="space-y-2">
-              <h3 className="font-semibold text-green-900">File Details:</h3>
-              <div className="space-y-2">
-                {result.files.map((file: any) => (
-                  <div key={file.filename} className="bg-white rounded p-3 flex justify-between items-center">
-                    <span className="font-medium">{file.filename}</span>
-                    <div className="flex gap-4 text-sm">
-                      <span className="text-green-700">{file.imported} imported</span>
-                      <span className="text-yellow-700">{file.duplicates} duplicates</span>
-                    </div>
-                  </div>
-                ))}
-              </div>
-            </div>
-          )}
-
-          {/* Cross-account warnings */}
-          {result.cross_account_warning_count > 0 && (
-            <div className="bg-yellow-50 border border-yellow-200 rounded-lg p-4 space-y-2">
-              <h3 className="font-semibold text-yellow-800">
-                Cross-Account Matches ({result.cross_account_warning_count})
-              </h3>
-              <p className="text-sm text-yellow-700">
-                These transactions appear to match existing transactions in other accounts.
-                This could indicate transfers between accounts.
-              </p>
-            </div>
-          )}
-        </div>
-      )}
+      {result && <ImportResult result={result} />}
     </div>
   )
 }
