+++ frontend/src/components/format-mapper/ConfidenceDisplay.tsx
@@ -1,0 +1,76 @@
+'use client'
+
+import { ColumnHint, getConfidenceColors } from '@/types/customFormat'
+
+// Icon component based on confidence level
+export function ConfidenceIcon({ confidence, size = 'sm' }: { confidence: number; size?: 'sm' | 'lg' }) {
+  const sizeClass = size === 'lg' ? 'w-5 h-5' : 'w-4 h-4'
+
+  if (confidence >= 0.8) {
+    return (
+      <svg className={`${sizeClass} text-green-500`} fill="currentColor" viewBox="0 0 20 20">
+        <path fillRule="evenodd" d="M10 18a8 8 0 100-16 8 8 0 000 16zm3.707-9.293a1 1 0 00-1.414-1.414L9 10.586 7.707 9.293a1 1 0 00-1.414 1.414l2 2a1 1 0 001.414 0l4-4z" clipRule="evenodd" />
+      </svg>
+    )
+  } else if (confidence >= 0.6) {
+    return (
+      <svg className={`${sizeClass} text-yellow-500`} fill="currentColor" viewBox="0 0 20 20">
+        <path fillRule="evenodd" d="M8.257 3.099c.765-1.36 2.722-1.36 3.486 0l5.58 9.92c.75 1.334-.213 2.98-1.742 2.98H4.42c-1.53 0-2.493-1.646-1.743-2.98l5.58-9.92zM11 13a1 1 0 11-2 0 1 1 0 012 0zm-1-8a1 1 0 00-1 1v3a1 1 0 002 0V6a1 1 0 00-1-1z" clipRule="evenodd" />
+      </svg>
+    )
+  }
+  return (
+    <svg className={`${sizeClass} text-red-500`} fill="currentColor" viewBox="0 0 20 20">
+      <path fillRule="evenodd" d="M10 18a8 8 0 100-16 8 8 0 000 16zM8.707 7.293a1 1 0 00-1.414 1.414L8.586 10l-1.293 1.293a1 1 0 101.414 1.414L10 11.414l1.293 1.293a1 1 0 001.414-1.414L11.414 10l1.293-1.293a1 1 0 00-1.414-1.414L10 8.586 8.707 7.293z" clipRule="evenodd" />
+    </svg>
+  )
+}
+
+// Row showing confidence for a detected column
+export function ColumnConfidenceRow({
+  label,
+  column,
+  hint,
+  isSet
+}: {
+  label: string
+  column: string
+  hint?: ColumnHint
+  isSet: boolean
+}) {
+  if (!isSet) {
+    return (
+      <div className="flex items-center gap-2 text-sm">
+        <svg className="w-4 h-4 text-red-500" fill="currentColor" viewBox="0 0 20 20">
+          <path fillRule="evenodd" d="M10 18a8 8 0 100-16 8 8 0 000 16zM8.707 7.293a1 1 0 00-1.414 1.414L8.586 10l-1.293 1.293a1 1 0 101.414 1.414L10 11.414l1.293 1.293a1 1 0 001.414-1.414L11.414 10l1.293-1.293a1 1 0 00-1.414-1.414L10 8.586 8.707 7.293z" clipRule="evenodd" />
+        </svg>
+        <span className="w-24 font-medium text-theme-muted">{label}:</span>
+        <span className="text-red-600 dark:text-red-400">Not detected</span>
+      </div>
+    )
+  }
+
+  const confidence = hint?.confidence ?? 0
+  const confidencePercent = Math.round(confidence * 100)
+  const { textColor, barColor } = getConfidenceColors(confidence)
+
+  return (
+    <div className="flex items-center gap-2 text-sm">
+      <ConfidenceIcon confidence={confidence} />
+      <span className="w-24 font-medium text-theme-muted">{label}:</span>
+      <span className="font-mono text-theme">{column}</span>
+      <div className="flex items-center gap-1 ml-2">
+        <div className="w-16 h-2 bg-gray-200 dark:bg-gray-700 rounded-full overflow-hidden">
+          <div
+            className={`h-full ${barColor} transition-all`}
+            style={{ width: `${confidencePercent}%` }}
+          />
+        </div>
+        <span className={`text-xs ${textColor}`}>{confidencePercent}%</span>
+      </div>
+      {hint?.format_display && (
+        <span className="text-xs text-theme-muted ml-1">({hint.format_display})</span>
+      )}
+    </div>
+  )
+}
